+++ functions.php
@@ -31,10 +31,25 @@ function time_to_end ($str_time) {
 function db_connect ($config) {
     $link = mysqli_connect($config["host"], $config["user"], $config["password"], $config["database"]);
     if ($link == false) {
-        print ("Ошибка подключения: " . mysqli_connect_error());
+        die ("Ошибка подключения: " . mysqli_connect_error());
     }
-    else {
-        mysqli_set_charset($link, "utf8");
-        return $link;
-    }
+    mysqli_set_charset($link, "utf8");
+    return $link;
+}
+
+function get_categories($link) {
+    $sql = "SELECT * FROM categories";
+    $result = mysqli_query($link, $sql);
+    $category = mysqli_fetch_all($result, MYSQLI_ASSOC);
+    return $category;
+}
+
+function get_lots($link) {
+    $sql = "SELECT l.title AS name, c.title AS category, l.price AS price, l.image, l.end_time
+             FROM lots l
+             JOIN categories c ON c.id = l.category_id
+             WHERE l.winner_id IS NULL";
+    $result = mysqli_query($link, $sql);
+    $lots = mysqli_fetch_all($result, MYSQLI_ASSOC);
+    return $lots;
 }
