+++ mysql_helper.php
@@ -44,29 +44,3 @@ function db_get_prepare_stmt($link, $sql, $data = []) {
     return $stmt;
 }
 
-function get_categories($link) {
-    $result = [];
-    $sql = "SELECT * FROM categories";
-    $stmt = db_get_prepare_stmt($link, $sql);
-    mysqli_stmt_execute($stmt);
-    $res = mysqli_stmt_get_result($stmt);
-    if ($res) {
-        $result = mysqli_fetch_all($res, MYSQLI_ASSOC);
-    }
-    return $result;
-}
-
-function get_lots($link) {
-    $sql = "SELECT l.title AS name, c.title AS category, l.price AS price, l.image, l.end_time
-             FROM lots l
-             JOIN categories c ON c.id = l.category_id
-             WHERE l.winner_id IS NULL";
-    $result = [];
-    $stmt = db_get_prepare_stmt($link, $sql);
-    mysqli_stmt_execute($stmt);
-    $res = mysqli_stmt_get_result($stmt);
-    if ($res) {
-        $result = mysqli_fetch_all($res, MYSQLI_ASSOC);
-    }
-    return $result;
-}
